# GitMiner/Documentation/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /gitminer/commits/{id}:
    get:
      tags:
      - Commit
      - get
      - commits
      - id
      summary: Retrieve a commit by id
      description: Get a commit by its identifier
      operationId: findOne_3
      parameters:
      - name: id
        in: path
        description: Id of the commit to find
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commit'
        "404":
          description: Not Found
    put:
      tags:
      - Commit
      - commits
      - id
      - put
      summary: Update a commit
      description: Put operation to update the commit with the given id
      operationId: update_3
      parameters:
      - name: id
        in: path
        description: Id of the commit to be updated
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Commit'
        required: true
      responses:
        "400":
          description: Bad Request
        "204":
          description: No Content
        "404":
          description: Not Found
    delete:
      tags:
      - Commit
      - commits
      - id
      - delete
      summary: Update a commit
      description: Put operation to update the commit with the given id
      operationId: delete_3
      parameters:
      - name: id
        in: path
        description: Id of the commit to be deleted
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
        "204":
          description: No Content
        "404":
          description: Not Found
...
  /gitminer/commits:
    get:
      tags:
      - Commit
      - get
      - commits
      summary: Retrieve all commits
      description: Get al commits
      operationId: findCommits
      parameters:
      - name: email
        in: query
        description: email of the author of the comments to find
        required: false
        schema:
          type: string
      - name: committername
        in: query
        description: committer of the comments to find
        required: false
        schema:
          type: string
      - name: beforedate
        in: query
        description: limit date of the comments to be find
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: order
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: number of elements per page
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commit'
        "404":
          description: Not Found
    post:
      tags:
      - Commit
      - post
      - commits
      summary: Upload a commit
      description: Post operation to save a commit in the database
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Commit'
        required: true
      responses:
        "400":
          description: Bad Request
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commit'
...
components:
  schemas:
...
    Commit:
      required:
      - author_name
      - authored_date
      - committed_date
      - committer_name
      - web_url
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        author_name:
          type: string
        author_email:
          type: string
        authored_date:
          type: string
        committer_name:
          type: string
        committer_email:
          type: string
        committed_date:
          type: string
        web_url:
          type: string
...